Alomst Done
only without Auth,
and trans to be deleted when acc is deleted

+++ Assessments/Assignment 5/Banking App/app/src/main/java/com/cmps312/bankingapp/ui/account/AddUpdateAccountFragment.kt
@@ -32,7 +32,7 @@ class AddUpdateAccountFragment : Fragment(R.layout.fragment_add_update_account) 
             if (isValidForm()) {
                 accountViewModel.apply {
                     account.name = nameEdt.text.toString()
-                    account.accountNumber = accountNoSp.text.toString().toInt()
+                    account.accountNumber = accountNoSp.text.toString()
                     account.acctType = typeSpinner.selectedItem.toString()
                     account.balance = balanceEdt.text.toString().toDouble()
                 }
